fixed modbus protocol async read, memory reader

+++ src/FlomtManager.Core/Constants/DeviceConstants.cs
@@ -3,7 +3,7 @@
     public class DeviceConstants
     {
         public const int MEMORY_SIZE_BYTES = 64 * 1024;
-        public const ushort MEMORY_SIZE_REGISTERS = MEMORY_SIZE_BYTES / 2;
+        public const int MEMORY_SIZE_REGISTERS = MEMORY_SIZE_BYTES / 2;
 
         public const byte MAX_PARAMETER_COUNT = 22;
